requirements.txt
compressai
torch
torchvision
matplotlib
numpy
scikit-image
Pillow